requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   instructor
    #   langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==25.3.0
    # via aiohttp
cachetools==6.2.0
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   typer
    #   uvicorn
dataclasses-json==0.6.7
    # via langchain-community
diskcache==5.6.3
    # via instructor
distro==1.9.0
    # via openai
docstring-parser==0.17.0
    # via instructor
dotenv==0.9.9
    # via pyphp (pyproject.toml)
exceptiongroup==1.3.0
    # via anyio
fastapi==0.118.0
    # via pyphp (pyproject.toml)
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.184.0
    # via google-generativeai
google-auth==2.41.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-genai==1.40.0
    # via instructor
google-generativeai==0.8.5
    # via pyphp (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.75.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   langsmith
    #   openai
httpx-sse==0.4.1
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
instructor==1.11.3
    # via pyphp (pyproject.toml)
jinja2==3.1.6
    # via instructor
jiter==0.10.0
    # via
    #   instructor
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonref==1.1.0
    # via instructor
langchain==0.3.27
    # via
    #   pyphp (pyproject.toml)
    #   langchain-community
langchain-community==0.3.30
    # via pyphp (pyproject.toml)
langchain-core==0.3.77
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.3.11
    # via langchain
langsmith==0.4.31
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.6
    # via langchain-community
openai==1.109.1
    # via instructor
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via
    #   fastapi
    #   google-genai
    #   google-generativeai
    #   instructor
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via
    #   instructor
    #   pydantic
pydantic-settings==2.11.0
    # via langchain-community
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.1.1
    # via
    #   dotenv
    #   pydantic-settings
pyyaml==6.0.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
requests==2.32.5
    # via
    #   google-api-core
    #   google-genai
    #   instructor
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==14.1.0
    # via
    #   instructor
    #   typer
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-community
starlette==0.48.0
    # via fastapi
tenacity==9.1.2
    # via
    #   google-genai
    #   instructor
    #   langchain-community
    #   langchain-core
tqdm==4.67.1
    # via
    #   google-generativeai
    #   openai
typer==0.19.2
    # via instructor
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   google-genai
    #   google-generativeai
    #   grpcio
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspect
    #   typing-inspection
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.37.0
    # via pyphp (pyproject.toml)
websockets==15.0.1
    # via google-genai
yarl==1.20.1
    # via aiohttp
zstandard==0.25.0
    # via langsmith
